pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "sharingan"
version = "0.1.0"
description = "OpenCV for Semantic Video Understanding with Temporal Reasoning"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sharingan Contributors"}
]
keywords = ["video", "computer-vision", "vlm", "temporal-reasoning", "semantic-understanding"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "torch>=1.10.0",
    "torchvision>=0.11.0",
    "Pillow>=8.0.0",
    "tqdm>=4.60.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
full = [
    "transformers>=4.20.0",
    "faiss-cpu>=1.7.0",
    "decord>=0.6.0",
]
gpu = [
    "faiss-gpu>=1.7.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "mypy>=0.950",
    "flake8>=4.0.0",
]
docs = [
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/sharingan/sharingan"
Documentation = "https://sharingan.readthedocs.io"
Repository = "https://github.com/sharingan/sharingan"
Issues = "https://github.com/sharingan/sharingan/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=sharingan --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["sharingan"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
